Fix: blank buffer appears when opening a file
When opening a file under version controlled system, e2wm
shows *e2wm:blank* buffer in the main window.  This happens
because kill-buffer-hook is called when vc-follow-link kills
newly opened buffer (it tries to follow symbolic link).

Inside of e2wm:kill-buffer-hook, WINS is nil because the
buffer is not shown yet.  But BUFFERS is also nil because
this buffer is not registered in history.  Therefore,
ignoring both cases in COND is the right choice.  However,
it executed the second clause because MAIN-WNAME is non-nil.
This clause must be executed only when WINS has some windows
in it.  Thus, the check for WIN is added.